laatste tijdsblok liefst niet #17,

+++ src/main/java/persistence/Marshalling.java
@@ -416,7 +416,7 @@ public class Marshalling {
 
 			for (int i = 0; i < sortedRingen.size(); i++) {
 				Ring currentRing = sortedRingen.get(i);
-				int tbRowBase = (int) (((float) i / 2) * (4 + ((float) Instellingen.Opties().TOTALETIJD / ringGroepDuur / 2)));
+				int tbRowBase = (int) ((i / 2) * (4 + (Instellingen.Opties().TOTALETIJD / ringGroepDuur / 2f)));
 				int tbColBase = (i % 2) * 5;
 
 				//Ringtitel
